kernel - pmap and vkernel work
* Remove the pmap.pm_token entirely.  The pmap is currently protected
  primarily by fine-grained locks and the vm_map lock.  The intention
  is to eventually be able to protect it without the vm_map lock at all.

* Enhance pv_entry acquisition (representing PTE locations) to include
  a placemarker facility for non-existant PTEs, allowing the PTE location
  to be locked whether a pv_entry exists for it or not.

* Fix dev_dmmap (struct dev_mmap) (for future use), it was returning a
  page index for physical memory as a 32-bit integer instead of a 64-bit
  integer.

* Use pmap_kextract() instead of pmap_extract() where appropriate.

* Put the token contention test back in kern_clock.c for real kernels
  so token contention shows up as sys% instead of idle%.

* Modify the pmap_extract() API to also return a locked pv_entry,
  and add pmap_extract_done() to release it.  Adjust users of
  pmap_extract().

* Change madvise/mcontrol MADV_INVAL (used primarily by the vkernel)
  to use a shared vm_map lock instead of an exclusive lock.  This
  significantly improves the vkernel's performance and significantly
  reduces stalls and glitches when typing in one under heavy loads.

* The new placemarkers also have the side effect of fixing several
  difficult-to-reproduce bugs in the pmap code, by ensuring that
  shared and unmanaged pages are properly locked whereas before only
  managed pages (with pv_entry's) were properly locked.

* Adjust the vkernel's pmap code to use atomic ops in numerous places.

* Rename the pmap_change_wiring() call to pmap_unwire().  The routine
  was only being used to unwire (and could only safely be called for
  unwiring anyway).  Remove the unused 'wired' and the 'entry'
  arguments.

  Also change how pmap_unwire() works to remove a small race condition.

* Fix race conditions in the vmspace_*() system calls which could lead
  to pmap corruption.  Note that the vkernel did not trigger any of
  these conditions, I found them while looking for another bug.

* Add missing maptypes to procfs's /proc/*/map report.

+++ sys/sys/device.h
@@ -122,7 +122,7 @@ struct dev_mmap_args {
 	struct dev_generic_args a_head;
 	vm_offset_t	a_offset;
 	int		a_nprot;
-	int		a_result;	/* page number */
+	int64_t		a_result;	/* page number */
 	struct file	*a_fp;
 };
 
@@ -330,7 +330,7 @@ int64_t dev_dpsize(cdev_t dev);
 int dev_dread(cdev_t dev, struct uio *uio, int ioflag, struct file *fp);
 int dev_dwrite(cdev_t dev, struct uio *uio, int ioflag, struct file *fp);
 int dev_dkqfilter(cdev_t dev, struct knote *kn, struct file *fp);
-int dev_dmmap(cdev_t dev, vm_offset_t offset, int nprot, struct file *fp);
+int64_t dev_dmmap(cdev_t dev, vm_offset_t offset, int nprot, struct file *fp);
 int dev_dmmap_single(cdev_t dev, vm_ooffset_t *offset, vm_size_t size,
 			struct vm_object **object, int nprot, struct file *fp);
 int dev_dclone(cdev_t dev);
